ADD selecting specific points on image AND saving points and person rectangle to variable

+++ src/mw/postureannotationtool/ui/model/Person.java
@@ -74,7 +74,7 @@ public class Person {
         points.add(point);
     }
 
-    public void setPersonRect(int x1, int y1, int x2, int y2) {
+    public void setPersonRect(double x1, double y1, double x2, double y2) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
